2_Read_Data - Added View Items button functions on App & Landing. Changed name from ItemsListComponent to ItemsList. setting.json updateBuildConfiguration & nullAnalysis. Formatted all files to 2 spaces - kg
