+++ front-end/src/Services/ItemServices.js
@@ -1,9 +1,9 @@
 import axios from 'axios';
 
-const BASE_URL = "http://http://localhost:8080/";
+const BASE_URL = "http://localhost:8080/";
 
 class ItemService {
     getAllItems(){
-        return axios.get('$BASE_URL)/items);
+        return axios.get('$BASE_URL)/items');
     }
 }
