--- front-end/src/components/screens/ItemsList.js
@@ -1,11 +1,0 @@
-import React from "react";
-import ItemsListComponent from "../layout/ItemsListComponent";
-
-export default function ItemsList() {
-  return (
-    <>
-      <h1 className="m-5 text-center">Items List</h1>
-      <ItemsListComponent />
-    </>
-  )
-}
